Add test for connection aliveness and remove duplicate from test_connection_manager
